requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
sqlalchemy
python-dotenv
charli3-dendrite
